Fix problem with poster path, start to implement ViewModel - this code is commented out for now
Problem was that getPosterPath() returned the value of poster path with “https://image.tmdb.org/t/p/w185” prefix, this means that it was saved into the database with the prefix.
When it was loaded again, the field posterPath already contained the prefix and it got the prefix added again.
I took away URL prefix from the getter and just made it a constant in MainActivity (outside of Movie.java).